requirements.txt
# Client (GUI)
PySide6

# Server (API)
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.3.0
SQLAlchemy>=2.0
alembic
pyodbc
python-jose[cryptography]
requests
